+++ pkg/providers/cps/resource_akamai_cps_third_party_enrollment.go
@@ -12,6 +12,7 @@ import (
 	"github.com/akamai/terraform-provider-akamai/v4/pkg/akamai"
 	"github.com/akamai/terraform-provider-akamai/v4/pkg/common/tf"
 	cpstools "github.com/akamai/terraform-provider-akamai/v4/pkg/providers/cps/tools"
+	"github.com/akamai/terraform-provider-akamai/v4/pkg/tools"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/customdiff"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
@@ -193,6 +194,11 @@ func resourceCPSThirdPartyEnrollmentCreate(ctx context.Context, d *schema.Resour
 		return diag.FromErr(err)
 	}
 
+	// save ClientMutualAuthentication and unset it in enrollment request struct
+	// create request must not have it set; in case its not nil, we will run update later to add it
+	clientMutualAuthentication := enrollment.NetworkConfiguration.ClientMutualAuthentication
+	enrollment.NetworkConfiguration.ClientMutualAuthentication = nil
+
 	req := cps.CreateEnrollmentRequest{
 		Enrollment:       *enrollment,
 		ContractID:       strings.TrimPrefix(contractID, "ctr_"),
@@ -203,6 +209,21 @@ func resourceCPSThirdPartyEnrollmentCreate(ctx context.Context, d *schema.Resour
 		return diag.FromErr(err)
 	}
 	d.SetId(strconv.Itoa(res.ID))
+
+	// when clientMutualAuthentication was provided, insert it back to enrollment and send the update request
+	if clientMutualAuthentication != nil {
+		logger.Debug("Updating ClientMutualAuthentication configuration")
+		enrollment.NetworkConfiguration.ClientMutualAuthentication = clientMutualAuthentication
+		req := cps.UpdateEnrollmentRequest{
+			EnrollmentID:              res.ID,
+			Enrollment:                *enrollment,
+			AllowCancelPendingChanges: tools.BoolPtr(true),
+		}
+		_, err := client.UpdateEnrollment(ctx, req)
+		if err != nil {
+			return diag.FromErr(err)
+		}
+	}
 
 	acknowledgeWarnings, err := tf.GetBoolValue("acknowledge_pre_verification_warnings", d)
 	if err != nil && !errors.Is(err, tf.ErrNotFound) {